requirements.txt
coinbase-advanced-py
requests
psycopg2-binary
